# faster convolve in LDM+Voigt mode  (2-30x faster on test cases)