<!-- Please be sure to check out our contributing guidelines,
https://github.com/radis/radis/blob/develop/CONTRIBUTING.md . -->

### Description
<!-- Provide a general description of what your pull request does. -->

Replacing numerical integral np.convolve with overlap-add method as in #220  
This applies only to the LDM+Voigt approach, which is not the current default (default is LDM+FFT), but which may be used in a near-future because it has built-in truncation (to better represent physical-truncations)


### Test and improvement

1.  First with a small broadening kernel (10 cm-1) on a large spectrum range (9000 cm-1)

New :

```python

from radis import calc_spectrum


s = calc_spectrum(1000, 10000, molecule='CO',
              Tgas=3000,
              databank='hitemp',
              broadening_method='voigt',
              broadening_max_width=10,
              cutoff=0,
              verbose=3)

#%% With Oaconvolve
"""
Scaling equilibrium linestrength
0.05s - Scaled equilibrium linestrength
0.01s - Calculated lineshift
d:\github\radis\radis\misc\warning.py:341: MissingSelfBroadeningWarning: Self-broadening temperature coefficient Tdpsel not given in database: used Tdpair instead
  warnings.warn(WarningType(message))
d:\github\radis\radis\misc\warning.py:341: AccuracyWarning: Some lines are too narrow (FWHM ~ 0.02 cm⁻¹) for the current spectral grid (wstep=0.01). Please reduce wstep to below 0.0067 cm⁻¹. You can use wstep='auto' to get the optimal spectral grid value. You can also ignore by setting `warnings={'AccuracyWarning':'ignore'}` (if you know what you're doing!)
  warnings.warn(WarningType(message))
0.11s - Calculate broadening HWHM
Calculating line broadening (437240 lines: expect ~ 43.77s on 1 CPU)
...... 0.02s - Precomputed DLM lineshapes (76)
...... 0.00s - Initialized vectors
...... 0.28s - Get closest matching line & fraction
...... 0.38s - Distribute lines over DLM
...... 5.74s - Convolve and sum on spectral range           👈👈👈
6.49s - Calculated line broadening
0.14s - Calculated other spectral quantities
6.89s - Spectrum calculated (before object generation)
0.03s - Generated Spectrum object
6.92s - Spectrum calculated
"""
s.store("test_oaconvolve.spec")
```

Running the same before the change :
```python
#%% With np.convolve
"""
Scaling equilibrium linestrength
0.05s - Scaled equilibrium linestrength
0.01s - Calculated lineshift
0.11s - Calculate broadening HWHM
d:\github\radis\radis\misc\warning.py:341: MissingSelfBroadeningWarning: Self-broadening temperature coefficient Tdpsel not given in database: used Tdpair instead
  warnings.warn(WarningType(message))
d:\github\radis\radis\misc\warning.py:341: AccuracyWarning: Some lines are too narrow (FWHM ~ 0.02 cm⁻¹) for the current spectral grid (wstep=0.01). Please reduce wstep to below 0.0067 cm⁻¹. You can use wstep='auto' to get the optimal spectral grid value. You can also ignore by setting `warnings={'AccuracyWarning':'ignore'}` (if you know what you're doing!)
  warnings.warn(WarningType(message))
Calculating line broadening (437240 lines: expect ~ 43.77s on 1 CPU)
...... 0.02s - Precomputed DLM lineshapes (76)
...... 0.00s - Initialized vectors
...... 0.28s - Get closest matching line & fraction
...... 0.35s - Distribute lines over DLM
...... 12.23s - Convolve and sum on spectral range           👈👈👈
12.94s - Calculated line broadening
0.15s - Calculated other spectral quantities
13.34s - Spectrum calculated (before object generation)
0.03s - Generated Spectrum object
13.37s - Spectrum calculated
"""
s.store("test_npconvolve.spec")

#%% Compare
from radis import plot_diff, load_spec
plot_diff(load_spec("test_npconvolve.spec"), load_spec("test_oaconvolve.spec"))
```
![image](https://user-images.githubusercontent.com/16088743/126902112-210b446f-adfa-41a5-943c-04f16cc22464.png)


👉 So about 2x faster on test case (broadening_width is 10 cm-1) , and same output (diff is 1e-15 above)




2.  Now with a larger broadening kernel (100 cm-1) on the same large spectrum range (9000 cm-1)

New :

```python

s = calc_spectrum(1000, 10000, molecule='CO',
              Tgas=3000,
              databank='hitemp',
              broadening_method='voigt',
              broadening_max_width=100,
              cutoff=0,
              verbose=3)

#%% With Oaconvolve
"""
Scaling equilibrium linestrength
0.08s - Scaled equilibrium linestrength
0.01s - Calculated lineshift
d:\github\radis\radis\misc\warning.py:319: MissingSelfBroadeningWarning: Self-broadening temperature coefficient Tdpsel not given in database: used Tdpair instead
  
0.14s - Calculate broadening HWHM
Calculating line broadening (440563 lines: expect ~ 440.61s on 1 CPU)
...... 0.14s - Precomputed DLM lineshapes (76)
...... 0.00s - Initialized vectors
...... 0.30s - Get closest matching line & fraction
...... 0.37s - Distribute lines over DLM
...... 8.49s - Convolve and sum on spectral range           👈👈👈
9.37s - Calculated line broadening
0.16s - Calculated other spectral quantities
9.85s - Spectrum calculated (before object generation)
0.03s - Generated Spectrum object
9.88s - Spectrum calculated
"""
s.store("test_oaconvolve_100")
```

```python
#%% With np.convolve
"""
Scaling equilibrium linestrength
0.06s - Scaled equilibrium linestrength
0.01s - Calculated lineshift
d:\github\radis\radis\misc\warning.py:319: MissingSelfBroadeningWarning: Self-broadening temperature coefficient Tdpsel not given in database: used Tdpair instead
  
d:\github\radis\radis\misc\warning.py:319: AccuracyWarning: Some lines are too narrow (FWHM ~ 0.02 cm⁻¹) for the current spectral grid (wstep=0.01). Please reduce wstep to below 0.0066 cm⁻¹. You can use wstep='auto' to get the optimal spectral grid value. You can also ignore by setting `warnings={'AccuracyWarning':'ignore'}` (if you know what you're doing!)
  
0.11s - Calculate broadening HWHM
Calculating line broadening (440563 lines: expect ~ 440.61s on 1 CPU)
...... 0.14s - Precomputed DLM lineshapes (76)
...... 0.00s - Initialized vectors
...... 0.29s - Get closest matching line & fraction
...... 0.41s - Distribute lines over DLM
...... 267.53s - Convolve and sum on spectral range           👈👈👈
268.46s - Calculated line broadening
0.16s - Calculated other spectral quantities
268.88s - Spectrum calculated (before object generation)
0.03s - Generated Spectrum object
268.90s - Spectrum calculated
"""
s.store("test_npconvolve_100")

#%% Compare
from radis import plot_diff, load_spec
plot_diff(load_spec("test_npconvolve_100.spec"), load_spec("test_oaconvolve_100.spec"))
```

![image](https://user-images.githubusercontent.com/16088743/126902481-bd678f7f-ff55-4162-9f61-0260f0bec463.png)


👉 This time it's 26x faster (broadening_width is 100 cm-1) , and same output (diff is 1e-15 above)

So significant improvement for large broadening truncations ; which is eventually what we want. 



<!-- If the pull request closes any open issues you can add this.
If you replace <Issue Number> with a number, GitHub will automatically link it.
If this pull request is unrelated to any issues, please remove
the following line. -->