@anandxkumar 2021-07-25
That's a very significant performance boost! It might just be on par with FFT now!